refactor tournament generation script
